fscryptctl: upload of mbedTLS sha512 code
This CL just uploads ARM's mbedTLS (formally PolarSSL) implementation
of the SHA512 hash. This will later be used by fscryptctl to generate
identifiers. The primary reason for including this implementation is so
fscryptctl doesn't need any external dependencies. This code will be
modified in a later commit to better suit the needs of fscryptctl.

Code was copied in from here: https://github.com/ARMmbed/mbedtls
Code is licensed under Apache 2, so we will not have compatibility
issues when releasing all of fscryptctl as Apache 2.