fscryptctl: get and set encryption policies
This commit adds in the ability for fscryptctl to setup encryption for a
directory or get the identifier of the current encryption policy for a
file or directory. This required adding policies, policy options, and
policy functions to libfscrypt/metadata/policyctl.

This commit also adds in the needed functions for translating between
identifiers and fscrypt_policies. Documentation for the added commands
can be seen by running "fscryptctl --help".

The README has been updated to document the additional functionality and
to show examples of using fscryptctl.