fscryptctl: insert a key into the keyring
This commit adds in the ability for fscryptctl to add a key to the
current session keyring. This is done via the "insert_key" command. This
takes a key from standard input, inserts it into the keyring with the
appropriate identifier, and outputs the corresponding identifier.

This commit also adds in constants and functions from "keyutils.h" and
changes key cleanup code to be uniform across the functions.

This change also updates the README for information on building and
using fscryptctl. Run "fscryptctl --help" to see full flag info.