fscryptctl: simplify SHA512 interface
This CL simplifies the mbedTLS SHA512 implementation. The interface to
the function now resembles the openssl interface. This commit also
removes any code related to SHA384 and formats the file with clang
format (Google style).

The code also simplifies the secure_wipe function and exposes it in the
header file.