fscryptctl: initial fscryptctl with get_descriptor
Fscryptctl is a small tool that provides some of the raw functionality
of filesystem encryption. It is designed to have no external library
dependencies and only performs one function currently: it takes a key
from standard input and outputs the corresponding descriptor. Having a
small tool for filesystem with minimal dependencies is useful for
embedded systems or for inclusion in a system package like util-linux.

This commit also adds in a Makefile and some <linux/fs.h> constants for
identifier size and key length. The build instructions and dependencies
are described in the README.