fscryptctl: added tests and improved errors
This commit adds in tests for the fscryptctl binary, using the python
testing framework pytest. Instructions for getting the necessary
dependencies and running the tests are in the README and CONTRIBUTING
files.

This commit also updates the error strings in fscryptctl to give more
descriptive error messages when getting or setting a policy fails.