Fix tests and doc to not use 'weak' AES-XTS keys
The current test code and example in README.md fails (at least with
Kernel 4.13 and likely earlier versions too) since the 512-bit keys used
are considered invalid by the Kernel's XTS implementation.

The AES-XTS key is split into two equal sized keys where the first key
is used as the actual data encryption key and the other half is used
for tweak construction. The in-kernel XTS implementation is written in
accord with IEEE P1619/D16 which requires the keys for data encryption
and tweak construction to be separate.

Fix this by using different non-weak keys.

Signed-off-by: David Gstir <david@sigma-star.at>