Rewrite the README file
Rewrite the README to get it fully up-to-date.

Describe the purpose of fscryptctl more accurately.  Since it's an
advanced tool, generally prefer to refer to the kernel documentation or
fscrypt documentation instead of documenting everything yet again.

Document the new commands, briefly explain the difference between v1 and
v2 encryption policies insofar as fscryptctl is concerned, and update
the examples to show using a v2 policy rather than v1.