Mark get_descriptor and insert_key as deprecated
The get_descriptor and insert_key commands are only useful for v1
encryption policies, so mark them as deprecated, move them to a
dedicated section of the source file, and update the help output.

(The replacement for insert_key is add_key.  The replacement for
get_descriptor is also add_key, as it doesn't seem that useful to have
the key identifier computation as a separate command.  We could add a
compute_identifier command later if it's needed, though.)